add(Auction.sol) auction smart contract example

+++ Lottery/Lottery.sol
@@ -18,27 +18,39 @@ contract Lottery {
         players.push(payable(msg.sender));
     }
 
-    function getBalance() public view returns(uint) {
-        require(msg.sender == manager, "You are not allowed to see the balance.");
+    function getBalance() public view returns (uint256) {
+        require(
+            msg.sender == manager,
+            "You are not allowed to see the balance."
+        );
         return address(this).balance;
     }
 
-    function random() public view returns(uint){
-        return uint(keccak256(abi.encodePacked(block.difficulty, block.timestamp, players.length)));
+    function random() public view returns (uint256) {
+        return
+            uint256(
+                keccak256(
+                    abi.encodePacked(
+                        block.difficulty,
+                        block.timestamp,
+                        players.length
+                    )
+                )
+            );
     }
 
     function pickWinner() public {
         require(msg.sender == manager);
         require(players.length >= 10);
 
-        uint r = random();
+        uint256 r = random();
         address payable winner;
-        uint index = r % players.length;
+        uint256 index = r % players.length;
 
         winner = players[index];
 
-        uint managerFee = (getBalance() * 10 ) / 100; // manager fee is 10%
-        uint winnerPrize = (getBalance() * 90 ) / 100;     // winner prize is 90%
+        uint256 managerFee = (getBalance() * 10) / 100; // manager fee is 10%
+        uint256 winnerPrize = (getBalance() * 90) / 100; // winner prize is 90%
 
         // transferring 90% of contract's balance to the winner
         winner.transfer(winnerPrize);
@@ -49,5 +61,4 @@ contract Lottery {
         // resetting the lottery for the next round
         players = new address payable[](0);
     }
-
 }
